whatsapp-python-bridge/requirements.txt
fastapi>=0.110
uvicorn>=0.23
requests>=2.31
pydantic>=2.7
python-dotenv>=1.0
pytest>=8.2
httpx>=0.27
